+++ tests/Tests/Parametizer/Subcommand/scripts/error-final-commit-forgotten-in-subcommand.php
@@ -10,7 +10,7 @@ require_once __DIR__ . '/../../../init-console.php';
  * We want to make sure all subcommand switches are committed when the script is executed.
  *
  * We set up a situation here where an error will be caused by recursive handling of all config branches.
- * The exception will happen for the config in which only one subcommand is set up, which is an illegal state.
+ * The exception will happen for the config that contains 2 subcommands with the same name, which is an illegal state.
  * If we get the error in the test without passing any parameter (without actually reaching the needed subcommand),
  * this means that the recursive handling actually took place.
  */
@@ -30,7 +30,8 @@ Parametizer::newConfig(throwOnException: true)
                 Parametizer::newConfig(throwOnException: true)
                     ->newSubcommandSwitch('switchme-l3')
                     ->newSubcommand('test31', Parametizer::newConfig(throwOnException: true))
-                    // No 2nd subcommand here, exception will be thrown.
+                    ->newSubcommand('test31', Parametizer::newConfig(throwOnException: true))
+                    // Same subcommand name, exception will be thrown.
 
                     ->newOption('--name-l3'),
             )