+++ tools/cli-toolkit/init.php
@@ -8,5 +8,11 @@ require_once __DIR__ . '/Classes/AutoloadDetector.php';
 
 AutoloadDetector::detectAndRequire();
 
+use Composer\Autoload\ClassLoader;
+
+$composerLoader = new ClassLoader(__DIR__ . '/../../vendor');
+$composerLoader->addPsr4('MagicPush\\CliToolkit\\Tools\\CliToolkit\\', [__DIR__]);
+$composerLoader->register();
+
 mb_internal_encoding('UTF-8');
 setlocale(LC_ALL, 'en_US.UTF-8');